Move fold & fold_witness from LCCCS to Multifolding
Also abstract fold_witness to prevent code repetition for LCCCS & CCCS
witness.
